Handled adding nodes/bubbles that do not anchor to the table
JS handling for bubbles that just connect to other bubbles and not to
anything on the imported data table.
It now draws the bubble and the size will be such that it overlaps the
entire width of all nodes that it connects to.

We still need to handle the links to connect correctly.