Fix for WK-31: Model display for eqp is messed up
Handle drawing lines for cases where one of the nodes might be big and
the other small, and the line from the center of one to other actually
does not reach the other. In those cases, we calculate the overlapping
region between the nodes and then draw lines in the center of the
overlapping region.