WK-200 Prefix to ns map order wrong in KR2RMLMappingGenerator
This was broken interactively because when we were creating the
KR2RMLMapping and adding prefixes according to the OntologyManager’s
prefix map, which goes from namespace to prefix, we were inadvertently
flipping the prefix and namespaces around.  Had to then make sure we
were writing out the mapping’s namespaces correctly in
KR2RMLMappingWriter.
Also removed system.out.println in shorthandurigenerator